requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
boto3==1.26.121
botocore==1.29.121
click==8.1.3
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
Flask==2.3.1
Flask-SQLAlchemy==3.0.3
ipykernel==6.21.2
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jupyter_client==8.0.3
jupyter_core==5.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.0
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.11
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.2
